Backend/requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.4
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PGPy==0.6.0
pillow==12.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
PyJWT==2.11.0
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
